requirements.txt
-i https://pypi.org/simple
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
idna==2.8
importlib-metadata==0.18
joblib==0.13.2
more-itertools==7.1.0
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
python-forecastio==1.4.0
pytz==2019.1
requests==2.22.0
responses==0.10.6
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
